requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
google_api_python_client==2.32.0
google_auth_oauthlib==0.4.6
mediapipe==0.8.9
numpy==1.21.4
opencv_python==4.5.4.60
protobuf==3.19.1
